move classes into hailtop
